+++ app/index.html
@@ -1,6 +1,9 @@
 <html lang="en">
 <body>
-  <button id="joke">Get a Quote</button>
+  <button id="joke">Get a Dumb Joke</button>
+ <!--  <p id="ajax-text"></p> -->
+
+  <button id="quote">Get a Quote</button>
   <p id="ajax-text"></p>
 
   <p>Piglatinify your name now!</p>
