+++ server.js
@@ -3,6 +3,8 @@ var app = express();
 var bodyparser = require("body-parser");
 var port = process.env.PORT || 3000;
 var piglatinify = require("./lib/piglatinify.js");
+var jokeLib = require("./lib/jokes.js");
+var quoteLib = require("./lib/quote.js");
 
 app.use(bodyparser.json());
 app.use(bodyparser.urlencoded({extended: true}));
@@ -23,30 +25,18 @@ app.post("/piglatin", function(req, res) {
  res.json(piglatined);
 });
 
-var quotes = [
- "I have not failed. I've just found 10,000 ways that won't work. - Thomas Edison.",
- "No matter where you go, there you are. ",
- "If it is a good idea, go ahead and do it. It is much easier to apologize than to get permission. - Rear Admiral Grace Hopper, USN, Ph.D"
-];
 
 app.get("/quote", function (req, res){
- var randomIndex = Math.floor(Math.random()*quotes.length);
- res.send(quotes[randomIndex]);
+ // var randomIndex = Math.floor(Math.random()*quotes.length);
+ var quoteSend = quoteLib();
+ res.send(quoteSend);
 });
 
 
-var jokes = [
-{ setup: "what are the sexiest animals in the barnyard?", 
-punchline: "Brown chicken brown cow!"},
-{ setup: "What was a jumpoline called after your mom jumped on it?", 
-punchline: "Trampoline"},
-{ setup: "how many tickles does it take to make an octopus laugh?",
-punchline: "ten tickles"}
-];
-
-app.get("/joke", function (req, res){
- var randomIndex = Math.floor(Math.random()*jokes.length);
- res.json(jokes[randomIndex]);
+app.get("/joke", function(req, res) {
+ // var randomIndex = Math.floor(Math.random()*jokes.length);
+ var jokeGen = jokeLib();
+ res.json(jokeGen);
 });
 
 app.listen(port, function(){
